Extract function createDebouncer() as a util
